% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{add.continuity}
\alias{add.continuity}
\title{Adds continuity correction to bugs data frame}
\usage{
add.continuity(
  bugs.df,
  add.continuity = list(),
  add.weight.continuity = list()
)
}
\arguments{
\item{bugs.df}{A data frame that must have data for binomial likelihod containing the
variables \code{"r"} and \code{"n"}}

\item{add.continuity}{Only applicable if \code{cleandat$likelihood=="binomial"}. A list of study identifiers
for which a continuity correction of 0.5 should be added. Values should match those in \code{cleandat$data.df$study_id}}

\item{add.weight.continuity}{Only applicable if \code{cleandat$likelihood=="binomial"}. A list of
study identifiers for which a continuity correction using the reciprocal opposite treatment arm
weight (Sweeting 2004) should be added. Values should match those in \code{cleandat$data.df$study_id}}
}
\description{
Adds continuity correction to bugs data frame
}
