% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{addbugsbits}
\alias{addbugsbits}
\title{Generates WinBUGS data and initial values}
\usage{
addbugsbits(
  bugs.df,
  bugs.trt,
  d.init.na = NULL,
  sd2.init.na = NULL,
  ume = FALSE,
  bugs.df.2 = NULL,
  bugs.binom = NULL,
  dinit = FALSE,
  sd2init = FALSE,
  deltainit = FALSE,
  covariate = FALSE,
  bias.adjust = FALSE,
  stroutput = FALSE
)
}
\description{
Generates WinBUGS data and initial values
}
