% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{bugstoexcel}
\alias{bugstoexcel}
\title{Outputs results from a BUGS model into an Excel workbook}
\usage{
bugstoexcel(
  outcome = "SMD_severe",
  bugsdat = NULL,
  modnam = "RE random class effect",
  res.format = "rds",
  decimals = 2,
  model.scale = NULL,
  addume = TRUE,
  ref = "Pill placebo",
  devplot = TRUE,
  netplot = TRUE,
  forestplot = TRUE,
  rankplot = FALSE,
  scalesd = NULL,
  node.names = NULL,
  ...
)
}
\arguments{
\item{outcome}{A string representing an outcome name. This is used to identify the
correct BUGS model, and to label the outputted Excel workbook.}

\item{bugsdat}{An object of class \code{"bugsdat"} containing the info used to run the BUGS
model (i.e. data, treatment codes, etc.)}

\item{modnam}{A string representing the name of the model used}

\item{res.format}{Can take either \code{"rds"} or \code{"xlsx"} to indicate whether the BUGS model
results should be accessed as an RDS or xlsx file (xlsx uses only posterior summary
statistics and so is faster).}

\item{decimals}{The number of decimal places to which to report numerical results}

\item{model.scale}{Indicates whether treatment effects are on \code{"log"} (e.g. OR, RoM) or
\code{"natural"} (e.g. mean difference) scale. If left as \code{NULL} (the default) this will be
inferred from the data.}

\item{addume}{Whether to add results for the Unrelated Mean Effects model alongside
direct estimates from the NMA. For these to be successfully read, a corresponding UME
BUGS model's results must be available and saved and an .RDS or .xlsx file (depending
on \code{res.format}) with the name \code{i.outcome}, where \code{outcome} is as specified in the
outcome parameter (e.g. \code{i.SMD_severe.xlsx}).}

\item{ref}{The name of the reference treatment against which the pooled treatment
effects will be plotted and shown in an \verb{Effect size} worksheet.}

\item{devplot}{Whether to add a Dev-Dev plot showing the deviance of the UME vs NMA
model. To allow this, \code{addume} must be set to \code{TRUE} (see requirements for this), and
the \code{dev} node of both NMA and UME models must have been monitored.}

\item{netplot}{Whether to plot a network plot within a \verb{Network Plots} worksheet}

\item{forestplot}{Whether to plot forest plots in the \verb{Effect size vs reference} worksheet}

\item{rankplot}{Whether to plot cumulative rankograms in the \code{Ranks} worksheet}

\item{scalesd}{A number indicating the SD to use for back-transforming SMD to a
particular measurement scale. If left as \code{NULL} then no transformation will be
performed (i.e. leave as \code{NULL} if not modelling SMDs)}

\item{node.names}{A list of names of parameters/nodes as named in the BUGS model. Each
element of the list should be a single character object with the name of the (monitored)
parameter in the BUGS model that can be read from the BUGS output. See details for info
on names, as well as default values.}
}
\description{
Outputs results from a BUGS model into an Excel workbook
}
\details{
For the outcome to be read in \code{"xlsx"} format the data in each \code{bugsresults.xlsx} tab (worksheet)
\strong{must} have the following named variables: \code{node}, \code{mean}, \code{sd}, \code{val2.5pc}, \verb{50\%}, \code{val97.5pc}, \code{pd}, \code{DIC}

Within \code{node.names}, element names in the list correspond to the following nodes:
\itemize{
\item \code{d="d"}: pooled relative treatment effects
\item \code{m="m"}: pooled relative class effects
\item \code{comps=NULL}: treatment-level mean differences / odds ratios between all treatments in the dataset.
The default is \code{NULL} in which case \code{bugstoexcel()} will search for nodes named either
\code{diff} or \code{or} and present mean differences or ORs accordingly.
\item \code{compsClass=NULL}: class-level mean differences / odds ratios between all classes in the dataset.
Default is \code{NULL} in which case \code{bugstoexcel()} will search for nodes named either
\code{ diffClass} or \code{orClass} and present mean differences or ORs accordingly.
\item \code{sd="sd"}: between-study SD
\item \code{sdClass="sd2"}: within-class SD
\item \code{rk="rk"}: treatment-level rankings
\item \code{rkClass="rkClass"}: class-level rankings
\item \verb{dev="dev}: residual deviance contributions
\item \code{totresdev="totresdev"}: total residual deviance
}
}
