% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{bugstopairs}
\alias{bugstopairs}
\title{Converts BUGS wide data to contrasts for each pair of treatments/classes in the dataset}
\usage{
bugstopairs(bugsdat, level = "treat")
}
\arguments{
\item{bugsdat}{An object of class \code{"bugsdat"} containing a BUGS-ready dataset in wide format}

\item{level}{Can take either \code{"treat"} or \code{"class"} to indicate that contrast info should be
given as pairs of treatments or classes respectively}
}
\description{
Converts BUGS wide data to contrasts for each pair of treatments/classes in the dataset
}
