% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{cleanbin}
\alias{cleanbin}
\title{Clean binary data}
\usage{
cleanbin(
  df,
  classcode,
  r = "discae",
  n = "nrand",
  connectonly = FALSE,
  drop.studies = NULL,
  drop.trt = NULL,
  connect.study = c(),
  ref = NULL
)
}
\arguments{
\item{df}{A cleaned data frame containing variables specified in y.b, y.f, y.cfb, sd.b, sd.f, sd.cfb, n.b and n.f.
Must also include the following variables:
\itemize{
\item \code{studyid}: study identifiers
\item \code{trt}: treatment names/identifiers corresponding to those in \code{classcode}
\item \code{class}: class names/identifiers corresponding to those in \code{classcode} (if classes are modelled)
}}

\item{classcode}{A data frame of treatment/class identifiers}

\item{r}{A string corresponding to the variable in \code{df} for the number of responders (numerator)}

\item{n}{A string corresponding to the variable in \code{df} for the total number of participants (denominator).
Can be either number randomised or number of completers depending on the desired analysis (ITT or PP)}

\item{connectonly}{Indicates whether treatments only used for connecting should be included (\code{TRUE}). Checks for
by excluding treatments based on \code{classcode}, specified as \code{classcode$connectonly==1}}

\item{drop.studies}{A vector of study names to be excluded from the analysis that
correspond to those given in \code{df$study_id}}

\item{connect.study}{A character vector of study identifiers matching those in \code{df$study_id}
that should be included specifically for network connectivity. If a study in \code{connect.study}
includes a treatment specified as \code{classcode$connectonly==1} (and if \code{connectonly=TRUE}) then
the study will be included (i.e. \code{connect.study} takes priority over \code{connectonly}).}

\item{ref}{An index for which treatment should become the reference treatment in \code{classcode}.}
}
\description{
Cleans binary data and generates treatment/class codes required for analysis.
}
