% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{cleancont.smd}
\alias{cleancont.smd}
\title{Clean continuous data for modelling using standardised mean differences (SMD)}
\usage{
cleancont.smd(
  df,
  classcode,
  y.b,
  y.f,
  y.cfb,
  sd.b,
  sd.f,
  sd.cfb,
  n.b,
  n.f,
  scale,
  predsd = TRUE,
  connectonly = FALSE,
  drop.studies = NULL,
  drop.trt = NULL,
  connect.study = c(),
  ref = NULL
)
}
\arguments{
\item{df}{A cleaned data frame containing variables specified in y.b, y.f, y.cfb, sd.b, sd.f, sd.cfb, n.b and n.f.
Must also include the following variables:
\itemize{
\item \code{studyid}: study identifiers
\item \code{trt}: treatment names/identifiers corresponding to those in \code{classcode}
\item \code{class}: class names/identifiers corresponding to those in \code{classcode} (if classes are modelled)
}}

\item{classcode}{A data frame of treatment/class identifiers}

\item{y.b}{A string corresponding to the variable in \code{df} for the mean at baseline}

\item{y.f}{A string corresponding to the variable in \code{df} for the mean at follow-up}

\item{y.cfb}{A string corresponding to the variable in \code{df} for the mean change from baseline}

\item{sd.b}{A string corresponding to the variable in \code{df} for the standard deviation at baseline}

\item{sd.f}{A string corresponding to the variable in \code{df} for the standard deviation at follow-up}

\item{sd.cfb}{A string corresponding to the variable in \code{df} for the standard deviation change from baseline}

\item{n.b}{A string corresponding to the variable in \code{df} for the number of participants randomised}

\item{n.f}{A string corresponding to the variable in \code{df} for the number of completers}

\item{scale}{A string corresponding to the variable in \code{df} that reports the measurement scale (e.g. HAM-D).
\code{cleancont.smd} assumes the use of SMDs is because multiple measurement scales are present. If only a single
scale is present then can create a variable in \code{df} that gives the same value for all rows of data.}

\item{predsd}{Indicates whether missing SDs should be predicted (\code{TRUE}) using other reported SDs (baseline,
follow-up and CFB) on the same measurement scale. If \code{TRUE}, plots and regression results for predictions
will be generated in addition to the cleaned data.}

\item{connectonly}{Indicates whether treatments only used for connecting should be included (\code{TRUE}). Checks for
by excluding treatments based on \code{classcode}, specified as \code{classcode$connectonly==1}}

\item{drop.studies}{A vector of study names to be excluded from the analysis that
correspond to those given in \code{df$study_id}}

\item{connect.study}{A character vector of study identifiers matching those in \code{df$study_id}
that should be included specifically for network connectivity. If a study in \code{connect.study}
includes a treatment specified as \code{classcode$connectonly==1} (and if \code{connectonly=TRUE}) then
the study will be included (i.e. \code{connect.study} takes priority over \code{connectonly}).}

\item{ref}{An index for which treatment should become the reference treatment in \code{classcode}.}
}
\description{
Cleans continuous data reported on multiple scales for modelling using SMDs. Also generates
treatment/class codes required for analysis.
}
