% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{codatocsv}
\alias{codatocsv}
\title{Extract coda from results and save to csv}
\usage{
codatocsv(
  outcome,
  ref = NULL,
  level = "class",
  writetocsv = TRUE,
  scalesd = NULL
)
}
\arguments{
\item{outcome}{A string representing an outcome name. This is used to identify the
correct BUGS model, and to label the outputted Excel workbook.}

\item{ref}{A treatment/class name, given as a string}

\item{writetocsv}{Can take \code{TRUE} for coda to be saved to csv, or \code{FALSE} for the coda
object to simply be refurned}

\item{scalesd}{A number indicating the value of a scale SD that should be used to convert SMDs back to
a specified scale}
}
\description{
Extract coda from results and save to csv
}
