% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{convert.numeric}
\alias{convert.numeric}
\title{Convert columns of a data frame to numeric (after checking removal of NAs)}
\usage{
convert.numeric(df, col, drop.ws = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{col}{A numeric vector representing column indices in \code{df} to convert to numeric}

\item{drop.ws}{Whether whitespaces should be dropped}
}
\description{
Convert columns of a data frame to numeric (after checking removal of NAs)
}
