% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{cumrank}
\alias{cumrank}
\title{Plots cumulative rankograms}
\usage{
cumrank(
  outcomes,
  level = "trt",
  byoutcome = TRUE,
  trtlist = list(),
  out.labs = outcomes
)
}
\arguments{
\item{outcomes}{A character vector of outcome names corresponding to worksheet names
in \code{"BUGSresults/bugsresults.xlsx"} \emph{or} Results workbooks in \code{"Results"} directory.
If multiple outcomes are specified then function plots a
separate panel for each treatment/class with multiple outcomes. Otherwise plots a single
outcome panel with all treatments/classes on the same plot.}

\item{level}{Can be \code{"trt"} for treatment-level results, or \code{"class"} for class-leve reuslts}

\item{byoutcome}{Indicates whether panels should be plotted by outcome (\code{TRUE}) or treatment/class}

\item{trtlist}{Used primarily within the \code{bugstoexcel()} to allow naming of treatments/classes.
A list in which each element is a set of treatment/class names for each outcome in \code{outcomes}. If left
as \code{NULL} (the default) then the function will load ranking results from outcome workbooks in \code{Results}
directory.}

\item{out.labs}{A character vector of labels for each of \code{outcomes}. Must be same length as \code{outcomes}}
}
\description{
Plots cumulative rankograms
}
