% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{devplot}
\alias{devplot}
\title{Plots a dev-dev plot}
\usage{
devplot(
  outcome,
  bugsdat = NULL,
  res.format = "xlsx",
  lim = NULL,
  add.modfit = TRUE,
  ratcut = 0,
  devcut = 5,
  direction = "nma",
  node.names = list(d = "d", m = "m", comps = NULL, compsClass = NULL, sd = "sd", sdClass
    = "sd2", rk = "rk", rkClass = "rkClass", dev = "dev", totresdev = "totresdev")
)
}
\arguments{
\item{outcome}{A character object specifying the name of the analysis for which to plot a dev-dev plot.
Must match a named worksheet in \code{"BUGSresults/bugsresults.xlsx"}, and there must be a corresponding UME model
that has been run and saved with the \code{i.} prefix. Note that \code{dev} nodes
\strong{must have been monitored and saved in both NMA and UME models to be able to plot a dev-dev plot}.}

\item{bugsdat}{An object of class \code{"bugsdat"} that includes the data used in the NMA. This will eventually
be deprecated to ensure it is stored elsewhere and doesn't need to be formally specified within this function.}

\item{res.format}{Can be \code{"xlsx"} or \code{"rds"} to indicate the format in which BUGS results have been saved.
The \code{savemodel()} function now saves results in both formats, and loading from xlsx is much more rapid so
leaving as the default is recommended.}

\item{lim}{A numeric vector of length 2 specifying the limits for the axes}

\item{add.modfit}{Indicates whether or not to include model fit statistics as annotations on the plot. Messy
but informative. Though the relevant info must have been monitored and saved.}

\item{ratcut}{A number representing the cutoff for the ratio of nmadev/umedev, above which (and if they also
exceed the value specified in \code{devcut}) points should be coloured red and their values returned in a
data frame.}

\item{devcut}{A number representing the cutoff value for the deviance, above which (and if they also
exceed the value specified in \code{ratcut}) points should be coloured red and their values returned in a data frame.}

\item{direction}{If given as \code{"nma"} it specifies \code{ratcut} should be calculated as nmadev/umedev and \code{devcut}
should be a threshold for the nma deviance. If given as \code{"ume"} it specifies \code{ratcut} should be calculated as
umedev/nmadev and \code{devcut} should be a threshold for the ume deviance.}

\item{node.names}{A list of names of parameters/nodes as named in the BUGS model. Each
element of the list should be a single character object with the name of the (monitored)
parameter in the BUGS model that can be read from the BUGS output. See details for info
on names, as well as default values.}
}
\description{
Plots a dev-dev plot
}
