% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{forestcols}
\alias{forestcols}
\title{Plot forest plot of results vs reference plus numerical results}
\usage{
forestcols(
  outcome,
  level = "treatment",
  writepdf = TRUE,
  width = 10,
  height = 12,
  exp = FALSE,
  cols = c("OR", "95\%CrI"),
  xlim = c(0.1, 0.5, 1, 5, 10, 25),
  refvalue = ifelse(exp == FALSE, 0, 1)
)
}
\arguments{
\item{outcome}{A string corresponding to the name of an outcome (as saved in a Results workbook)}

\item{level}{A string indicating results should be plotted at either the \code{"treatment"} or \code{"class"} level}

\item{writepdf}{A logical object indicating whether a pdf should be written to file}

\item{width}{Width of plot if \code{writepdf=TRUE}}

\item{height}{Height of plot if \code{writepdf=TRUE}}

\item{exp}{Logical object indicating whether treatment effects given in Results workbook should be exponentiated or not}

\item{cols}{A character vector of length 2 indicating the names of the columns for the tabular results}

\item{xlim}{A numeric vector indicating the ticks on the x axis}

\item{refvalue}{A number indicating where a vertical dashed line for the reference treatment effect should be drawn}
}
\description{
Plot forest plot of results vs reference plus numerical results
}
