% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{forestplot}
\alias{forestplot}
\title{Plots a basic forest plot for data frame of outcomes}
\usage{
forestplot(
  out.df,
  refline = TRUE,
  refvalue = 0,
  xlab = "Effect size",
  ylab = "Treatment/Class"
)
}
\description{
Use within internal \code{gtsu} functions, rather than by the user
}
