% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{forestsense}
\alias{forestsense}
\title{Plots results from multiple NMAs on same forest plot}
\usage{
forestsense(
  analyses = c("SMD_severe"),
  ref = "Pill placebo",
  refline = TRUE,
  param = "diff",
  analysis.lab = NULL,
  dodge = 0.5,
  caption = ""
)
}
\arguments{
\item{analyses}{A character vector of analyses to plot on the same forest plot. Elements must be the same
as names for workbooks in \code{"Results"} directory.}

\item{ref}{A named treatment/class to use as the reference (against which relative effects will be plotted).
Must be present in all analysis results given in \code{analyses} and must match the treatment/class level
indicated by \code{param}.}

\item{refline}{Whether to plot a vertical dashed line indicating the reference treatment effect}

\item{param}{A character object indicating the relative effect parameter from the NMA to be plotted.
Within this is indicated whether effects are plotted at class or treatment level. Typical options are \code{diff},
\code{diffClass}, \code{or}, \code{orClass}.}

\item{analysis.lab}{A character vector of the same length as \code{analyses} indicating the labels for each
analysis to be shown in the legend. Can use \verb{\\n} for a new line if the label should be on multiple lines.}

\item{dodge}{A number indicating how far apart to plot results from differnt analyses}

\item{caption}{A character object specifying a caption to include underneath the plot.}
}
\description{
Useful for showing sensitivity/subgroup anlaysis results
}
