% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{getmodfit}
\alias{getmodfit}
\title{Get model fit statistics from a BUGS model results}
\usage{
getmodfit(
  bugsmod,
  bugsdat,
  res.format = "rds",
  modnam = "NMA",
  decimals = 2,
  scalesd = NULL,
  node.names
)
}
\arguments{
\item{bugsmod}{A data frame of posterior summary statistics from a BUGS model.
Must include a column for pD and DIC (as well as a totresdev node).}

\item{bugsdat}{An object of class \code{"bugsdat"} containing the info used to run the BUGS
model (i.e. data, treatment codes, etc.)}

\item{res.format}{Can take either \code{"rds"} or \code{"xlsx"} to indicate whether the BUGS model
results should be accessed as an RDS or xlsx file (xlsx uses only posterior summary
statistics and so is faster).}

\item{modnam}{A string representing the name of the model used}

\item{decimals}{The number of decimal places to which to report numerical results}

\item{scalesd}{A number indicating the SD to use for back-transforming SMD to a
particular measurement scale. If left as \code{NULL} then no transformation will be
performed (i.e. leave as \code{NULL} if not modelling SMDs)}

\item{node.names}{A list of names of parameters/nodes as named in the BUGS model. Each
element of the list should be a single character object with the name of the (monitored)
parameter in the BUGS model that can be read from the BUGS output. See details for info
on names, as well as default values.}
}
\description{
Get model fit statistics from a BUGS model results
}
