% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{init.dir}
\alias{init.dir}
\title{Initialise directory}
\usage{
init.dir(wd = NULL)
}
\arguments{
\item{wd}{A path to the directory which will be the source for the analyses}
}
\description{
Creates relevant subfolders
}
