% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinma.functions.R
\name{multinmatoexcel}
\alias{multinmatoexcel}
\title{Outputs results from a \code{multinma} NMA model into an Excel workbook}
\usage{
multinmatoexcel(
  nma,
  ume = NULL,
  filenam,
  outcome = "Outcome",
  modnam = "RE model",
  decimals = 2,
  eform = FALSE,
  scalesd = NULL,
  trt_ref = nma$network$treatments[1],
  lower_better = TRUE,
  treatments_to_rank = nma$network$treatments,
  classes_to_rank = nma$network$classes,
  devplot = TRUE,
  netplot = TRUE,
  forestplot = TRUE,
  rankplot = FALSE,
  pval = 0.05,
  addlogo = TRUE,
  ...
)
}
\arguments{
\item{nma}{A \code{multinma} consistency model of class \code{"stan_nma"}}

\item{ume}{A \code{multinma} UME model of class \code{"stan_nma"}. Only needs to be specified
if user wants to add results for the Unrelated Mean Effects model alongside direct
estimates from the NMA.}

\item{filenam}{A string representing a file path and name. Must end in \code{.xlsx}}

\item{outcome}{A string representing an outcome name. This is used to
label the outputted Excel workbook.}

\item{modnam}{A string representing the name of the model used}

\item{decimals}{The number of decimal places to which to report numerical results}

\item{eform}{Indicates whether treatment effect outputs should be exponentiated (\code{TRUE}) or not
(i.e. if model was on log scale).}

\item{scalesd}{A number indicating the SD to use for back-transforming SMD to a
particular measurement scale. If left as \code{NULL} then no transformation will be
performed (i.e. leave as \code{NULL} if not modelling SMDs)}

\item{trt_ref}{The name of the reference treatment against which the pooled treatment
effects will be plotted and shown in an \verb{Effect size} worksheet.}

\item{lower_better}{Logical to indicate whether lower scores are ranked as better (\code{TRUE})
or not (\code{FALSE})}

\item{treatments_to_rank}{A character vector of treatment names (must be a subset of those
in \code{nma$network$treatments}) to include in rankings. Allows a decision subset treatments in the
network to be ranked.}

\item{classes_to_rank}{A character vector of class names (must be a subset of those in
\code{nma$network$treatments}) to include in rankings.}

\item{devplot}{Whether to add a Dev-Dev plot showing the deviance of the UME vs NMA
model. To allow this, \code{addume} must be set to \code{TRUE} (see requirements for this), and
the \code{dev} node of both NMA and UME models must have been monitored.}

\item{netplot}{Whether to plot a network plot within a \verb{Network Plots} worksheet}

\item{forestplot}{Whether to plot forest plots in the \verb{Effect size vs reference} worksheet}

\item{rankplot}{Whether to plot cumulative rankograms in the \code{Ranks} worksheet}

\item{pval}{Numeric threshold for p-value for direct vs indirect estimates. Note that this is
an approximation as it assumes the posterior is normally distributed, and the data is correlated
...i.e. it is not a true indicator of direct vs indirect (nodesplitting would be needed for that).
It should only be used to highlight comparisons for further investigation.}

\item{addlogo}{Logical to indicate if GTSU logo should be added}

\item{...}{Arguments to be sent to other \code{multinma} function}
}
\description{
Outputs results from a \code{multinma} NMA model into an Excel workbook
}
