% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{neatcri}
\alias{neatcri}
\title{Returns the posterior median and upper/lower 95\%CrI as a string}
\usage{
neatcri(x, decimals = 2, invert = FALSE)
}
\arguments{
\item{x}{A data frame containing variables for the lower 95\%CrI, median, and upper
95\%CrI of the posterior for each node}

\item{invert}{Whether to invert the upper and lower 95\%CrI (logical)}
}
\description{
Returns the posterior median and upper/lower 95\%CrI as a string
}
