% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{neatcri2}
\alias{neatcri2}
\title{Returns the posterior median and upper/lower 95\%CrI as a string \strong{based on multinma output}}
\usage{
neatcri2(x, colnams = c("50\%", "2.5\%", "97.5\%"), decimals = 2)
}
\arguments{
\item{x}{A data frame containing variables for the lower 95\%CrI, median, and upper
95\%CrI of the posterior for each node}

\item{colnams}{A vector of column names corresponding to
c(\verb{point estimate}, \verb{lower interval}, \verb{upper interval})}
}
\description{
Returns the posterior median and upper/lower 95\%CrI as a string \strong{based on multinma output}
}
