% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{netplot}
\alias{netplot}
\title{Plot network plot}
\usage{
netplot(
  plotdat,
  classcode,
  v.color = "connect",
  level = "trt",
  remove.loops = TRUE,
  legend = TRUE,
  edge.scale = NULL,
  v.scale = NULL,
  vertex.label.dist = 0,
  shortnames = FALSE,
  ...
)
}
\arguments{
\item{plotdat}{A dataframe of contrasts containing \code{t1} and \code{t2} for treatment/class names/codes,
and \code{nr} representing the number of studies that include this contrast. Can easily be generated using
\code{bugstopairs()}.}

\item{classcode}{An object of class \code{"classcode"}, which is a dataframe of treatment and class names and codes}

\item{v.color}{Can take \code{"connect"} to colour only nodes that are connected to the network or \code{"class"} to colour by class}

\item{level}{Can take \code{"trt"} to plot nodes at the treatment level or \code{"class"} to plot nodes at the class level}

\item{remove.loops}{A logical object to indicate whether loops that compare within a node should be plotted}

\item{legend}{Indicates whether a legend of class names should be plotted or not...sometimes setting to \code{TRUE} can
lead to difficult sizing of the network plot}

\item{edge.scale}{A number by which to scale the connecting (edge) widths. Can leave as \code{NULL} for all edges to
be plotted with the same width}

\item{v.scale}{A number by which to scale the vertex (node) size. Can leave as \code{NULL} for all vertices to
be plotted with the same size}

\item{vertex.label.dist}{A number indicating how far vertex (node) labels should be plotted from the vertices (nodes)}

\item{shortnames}{Can be set to true to label vertices (nodes) using \code{shorttreatment} or \code{shortclass} variables in \code{classcodes}}

\item{...}{Arguments to be sent to \code{igraph}}
}
\description{
Takes a dataset of contrasts and generates a network plot
}
