% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{netpptx}
\alias{netpptx}
\title{Plots network plot to a powerpoint file (editable)}
\usage{
netpptx(
  bugsdat,
  vertex.label.cex = 0.2,
  t.v.scale = NULL,
  c.v.scale = NULL,
  file = "Plots/network_discany.pptx",
  title = "Discontinuation due to any reason"
)
}
\description{
Note that this actually isn't necessary to produce an editable plot since a plot can instead just be saved as
an .emf file (metafile format) (e.g. using \code{devEMF::emf()}), which when opened in powerpoint is editable.
}
