% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{poolsd}
\alias{poolsd}
\title{Pools standard deviations (SD) weighted by \code{n}}
\usage{
poolsd(n, sd)
}
\arguments{
\item{n}{A numeric vector of sample sizes}

\item{sd}{A numeric vector of SDs}
}
\description{
Pools standard deviations (SD) weighted by \code{n}
}
