% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{prepbugs}
\alias{prepbugs}
\title{Prepare data for OpenBUGS/WinBUGS}
\usage{
prepbugs(
  cleandat,
  varshare = c(),
  bias.adjust = FALSE,
  covariate = NULL,
  scalesd = TRUE,
  add.continuity = list(),
  add.weight.continuity = list(),
  sd2.init.na = NULL,
  d.init.na = NULL,
  dinit = FALSE,
  sd2init = FALSE,
  deltainit = FALSE,
  ume = FALSE,
  stroutput = FALSE
)
}
\arguments{
\item{cleandat}{A list of inputs from a function such as cleanbin() or cleancont(). List elements must
be named:
\itemize{
\item \code{"data.df"} An arm-based data frame of cleaned data
\item \code{"trtcodes"} A data frame of treatment/class codes specific to this outcome/analysis
\item \code{"likelihood"} A string indicating the likelihood of the data (\code{"binomial"} or \code{"continuous"})
}}

\item{varshare}{A numeric vector of codes to indicate which classes should share the same class variance
in a random class model. Must be same length as number of treatments, and codes must be the same within
a class. Codes must be sequentially numbered (starting at 1).}

\item{bias.adjust}{Indicates whether bias variables should be added to data (\code{TRUE}) or not (\code{FALSE})}

\item{covariate}{A string representing the name of a variable in \code{cleandat$data.df} to be included as
a covariate in a Network Meta-Regression. \code{NULL} indicates no covariate should be included.}

\item{scalesd}{Indicates whether the variable \code{scalesd} in \code{cleandat$data.df} should be used to
standardise SD (\code{TRUE}) rather than using the study-specific SD. E.g. If an internal/external reference SD
has been calculated for each scale, then the variable \code{scalesd} should be included in \code{cleandat$data.df}.
Values for this should be consistent within measurements from the same scale.}

\item{add.continuity}{Only applicable if \code{cleandat$likelihood=="binomial"}. A list of study identifiers
for which a continuity correction of 0.5 should be added. Values should match those in \code{cleandat$data.df$study_id}}

\item{add.weight.continuity}{Only applicable if \code{cleandat$likelihood=="binomial"}. A list of
study identifiers for which a continuity correction using the reciprocal opposite treatment arm
weight (Sweeting 2004) should be added. Values should match those in \code{cleandat$data.df$study_id}}

\item{sd2.init.na}{A numeric vector of class codes to indicate which classes do not need an initial value
for sd2 (the within-class SD) - i.e. that they have been modelled with a fixed class effect.}

\item{d.init.na}{A numeric vector of treatment codes to indicate which treatment effects do not need an initial
value - i.e. that they have been modelled with a fixed class effect.}

\item{dinit}{Indicates whether initial values for d should be added. If \code{FALSE} then this
suggests a fixed class effect should be modelled for all classes.}

\item{sd2init}{Indicates whether initial values for sd2 (within-class effect) should be added for any classes
or not. If \code{FALSE} then this indicates all classes should be modelled with a fixed class effect.}

\item{deltainit}{Indicates whether initial values should be added for deltainit. If \code{FALSE} then this
suggests a fixed treatment effect model, whereas \code{TRUE} suggests a random treatment effects model
will be fitted.}

\item{ume}{\code{TRUE} indicates that a UME model will be fitted and therefore the relevant initial values
will be generated.}

\item{stroutput}{If \code{TRUE} this indicates that data and initial values should be written to a .csv file
named \code{BUGSdata/bugsdata.csv} (i.e. stored in the \code{BUGSdata} directory). This can then be copied and pasted
directly into an OpenBUGS or WinBUGS file so that the model can be run manually by the user rather than
through \code{R2OpenBUGS}/\code{R2WinBUGS}.}
}
\value{
Returns an object of class \code{"bugsdat"} that contains all the info required for an analysis in
OpenBUGS or WinBUGS.

An list object of class \code{"bugsdat"} containing the following named elements:
\itemize{
\item \code{arraydat} : A wide (one row per study) data frame of BUGS data with the necessary
variables for analysis
\item \code{trtcodes} : A data frame of treatment names and indices matching the treatment codes
in \code{arraydat}. Can also include corresponding class names and codes for class effect
models.
\item \code{data.list} : Data for the model saved as a list of objects for the BUGS model in a
format that can be used directly with \code{R2OpenBUGS}/\code{R2WinBUGS}.
\item \code{inits} : A list of initial values in a format that can be used directly with
\code{R2OpenBUGS}/\code{R2WinBUGS}.
}
}
\description{
Prepare data for OpenBUGS/WinBUGS
}
