% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{rdstoxlsx}
\alias{rdstoxlsx}
\title{Writes summary BUGS data from RDS files of models to an xlsx file named "bugsresults.xlsx"}
\usage{
rdstoxlsx(outcomes = "dyspar", overwrite = TRUE, bugs.result = NULL)
}
\arguments{
\item{outcomes}{a vector of outcome names to write from rds to xlsx}

\item{overwrite}{indicates whether previous data should be overwritten or not}

\item{bugs.result}{Can be used to pass a data frame of posterior summary statistics from
a BUGS model directly to the function.}
}
\value{
Creates an excel file named "bugresults.xlsx" within the subdirectory BUGSresults
}
\description{
Writes summary BUGS data from RDS files of models to an xlsx file named "bugsresults.xlsx"
}
