% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.functions.R
\name{recode.trt}
\alias{recode.trt}
\title{Returns a data frame of new treatment and class codes with \code{ref} as the reference treatment}
\usage{
recode.trt(df, classcode, ref = NULL)
}
\description{
Returns a data frame of new treatment and class codes with \code{ref} as the reference treatment
}
