% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{savemodel}
\alias{savemodel}
\title{Save bugs models}
\usage{
savemodel(bugs.result, bugsdat, outcome, rds = FALSE, xlsx = TRUE)
}
\arguments{
\item{bugs.result}{An object generated from \code{R2OpenBUGS::bugs()}}

\item{bugsdat}{An object of class \code{"bugsdat"}}

\item{rds}{Whether to save a \code{.RDS} file of the model. If set to \code{TRUE} then
the function may take longer to run (BUGS files can be big since they contain
all the MCMC draws)}

\item{xlsx}{Whether to save a \code{.xlsx} file of posterior summary statistics}
}
\description{
Save bugs models
}
