% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.functions.R
\name{summary_multinma_matrix}
\alias{summary_multinma_matrix}
\title{Create posterior summaries for MCMC matrix}
\usage{
summary_multinma_matrix(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\description{
Create posterior summaries for MCMC matrix
}
